+++ packages/notification-services-controller/src/NotificationServicesController/processors/process-snap-notifications.ts
@@ -2,6 +2,7 @@ import { v4 as uuid } from 'uuid';
 
 import type { INotification } from '../types';
 import type { RawSnapNotification } from '../types/snaps';
+import { getNotificationSubtype } from '../utils/get-notification-subtype';
 
 /**
  * Processes a snap notification into a normalized shape.
@@ -15,6 +16,7 @@ export const processSnapNotification = (
   const { data, type, readDate } = snapNotification;
   return {
     id: uuid(),
+    notification_subtype: getNotificationSubtype(snapNotification),
     readDate,
     createdAt: new Date().toISOString(),
     isRead: false,